chore: move eventSortingConfigAtom into separate file
