+++ src/core/config/types.ts
@@ -1,4 +1,4 @@
-import type { SortConfig } from '~features/events_list/atoms/sortedEventList';
+import type { EventSortConfig } from '~features/events_list/atoms/eventSortingConfig';
 import type { UserDto } from '~core/app/user';
 import type { AppFeatureType } from '~core/app/types';
 import type { LayerDetailsDto } from '~core/logical_layers/types/source';
@@ -102,5 +102,5 @@ export type AboutFeatureConfig = {
 };
 
 export type EventsListFeatureConfig = {
-  initialSort?: SortConfig;
+  initialSort?: EventSortConfig;
 };
